+++ packages/web3-eth-contract/CHANGELOG.md
@@ -392,3 +392,4 @@ Documentation:
 
 -   `populateTransaction` was added to contract methods (#7124)
 
+-   Contract has `setTransactionMiddleware` and `getTransactionMiddleware` for automatically passing to `sentTransaction` for `deploy` and `send` functions (#7138)